Fix: Resolve build errors and improve stability
This commit addresses potential build errors and enhances the overall stability of the application.
