+++ src/components/connection/hooks/useConnectionDialog.ts
@@ -3,6 +3,7 @@ import { useState } from "react";
 import { Platform } from "@/types/platform";
 import { useToast } from "@/hooks/use-toast";
 import { isPlatformSupported } from "@/handlers/platformHandlers";
+import { useOAuthHandlers } from "./useOAuthHandlers";
 
 interface UseConnectionDialogProps {
   platform: Platform | null;
@@ -13,17 +14,15 @@ interface UseConnectionDialogProps {
 export function useConnectionDialog({ platform, onConnect, onClose }: UseConnectionDialogProps) {
   const [credentials, setCredentials] = useState<Record<string, string>>({});
   const [isConnecting, setIsConnecting] = useState(false);
-  const [isGitHubAuthing, setIsGitHubAuthing] = useState(false);
-  const [isTwitterAuthing, setIsTwitterAuthing] = useState(false);
-  const [isFacebookAuthing, setIsFacebookAuthing] = useState(false);
-  const [isInstagramAuthing, setIsInstagramAuthing] = useState(false);
-  const [isLinkedInAuthing, setIsLinkedInAuthing] = useState(false);
-  const [isTikTokAuthing, setIsTikTokAuthing] = useState(false);
-  const [isKooAuthing, setIsKooAuthing] = useState(false);
-  const [isShareChatAuthing, setIsShareChatAuthing] = useState(false);
   const { toast } = useToast();
 
   const isSupported = platform ? isPlatformSupported(platform.id) : false;
+
+  const oauthHandlers = useOAuthHandlers({
+    platform,
+    credentials,
+    onConnect,
+  });
 
   const handleConnect = async () => {
     if (!platform || !isSupported) {
@@ -62,219 +61,13 @@ export function useConnectionDialog({ platform, onConnect, onClose }: UseConnect
     setCredentials({});
   };
 
-  const handleGitHubOAuth = () => {
-    if (!platform) return;
-    
-    setIsGitHubAuthing(true);
-    try {
-      onConnect(platform.id, {});
-    } catch (err) {
-      setIsGitHubAuthing(false);
-      toast({
-        title: "GitHub OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleTwitterOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.clientId || !credentials.clientSecret) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your Twitter Client ID and Client Secret first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsTwitterAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsTwitterAuthing(false);
-      toast({
-        title: "Twitter OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleFacebookOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.appId || !credentials.appSecret) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your Facebook App ID and App Secret first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsFacebookAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsFacebookAuthing(false);
-      toast({
-        title: "Facebook OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleInstagramOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.appId || !credentials.appSecret) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your Facebook App ID and App Secret first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsInstagramAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsInstagramAuthing(false);
-      toast({
-        title: "Instagram OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleLinkedInOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.clientId || !credentials.clientSecret) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your LinkedIn Client ID and Client Secret first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsLinkedInAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsLinkedInAuthing(false);
-      toast({
-        title: "LinkedIn OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleTikTokOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.clientId || !credentials.clientSecret) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your TikTok Client ID and Client Secret first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsTikTokAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsTikTokAuthing(false);
-      toast({
-        title: "TikTok OAuth failed",
-        description: err instanceof Error ? err.message : "Could not start OAuth2.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleKooOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.email || !credentials.password) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your Koo email and password first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsKooAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsKooAuthing(false);
-      toast({
-        title: "Koo connection failed",
-        description: err instanceof Error ? err.message : "Could not connect to Koo.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleShareChatOAuth = () => {
-    if (!platform) return;
-    
-    if (!credentials.phone || !credentials.password) {
-      toast({
-        title: "Missing Credentials",
-        description: "Please enter your ShareChat phone and password first.",
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsShareChatAuthing(true);
-    try {
-      onConnect(platform.id, credentials);
-    } catch (err) {
-      setIsShareChatAuthing(false);
-      toast({
-        title: "ShareChat connection failed",
-        description: err instanceof Error ? err.message : "Could not connect to ShareChat.",
-        variant: "destructive",
-      });
-    }
-  };
-
   return {
     credentials,
     setCredentials,
     isConnecting,
-    isGitHubAuthing,
-    isTwitterAuthing,
-    isFacebookAuthing,
-    isInstagramAuthing,
-    isLinkedInAuthing,
-    isTikTokAuthing,
-    isKooAuthing,
-    isShareChatAuthing,
     isSupported,
     handleConnect,
     handleClose,
-    handleGitHubOAuth,
-    handleTwitterOAuth,
-    handleFacebookOAuth,
-    handleInstagramOAuth,
-    handleLinkedInOAuth,
-    handleTikTokOAuth,
-    handleKooOAuth,
-    handleShareChatOAuth,
+    ...oauthHandlers,
   };
 }
